+++ modules/gaming.nix
@@ -5,7 +5,7 @@
     services.xserver.inputClassSections = [ ''
       Identifier    "CTL-472"
       MatchProduct  "Wacom One"
-      Option        "TransformationMatrix" "0.4463039007311721 0 0.3095532576023331 0 0.9433962264150944 0.03966763902681231 0 0 1"
+      Option        "TransformationMatrix" "0.5381253317409767 0 0.02466001733346604 0 1.1986436400914755 0.11744785505874931 0 0 1"
     '' ];
 
     programs.steam.enable = true;
@@ -13,7 +13,7 @@
 
   hmModule = { pkgs, ... }: {
     home.packages = builtins.attrValues {
-      inherit (pkgs) polymc lutris;
+      inherit (pkgs) lutris prismlauncher;
     };
   };
 }