+++ overlays/firefox-addons/flake.nix
@@ -46,15 +46,6 @@
           meta = {};
         };
 
-        redirector = buildFirefoxXpiAddon {
-          pname = "redirector";
-          version = "3.5.3";
-          addonId = "redirector@einaregilsson.com";
-          url = "https://addons.mozilla.org/firefox/downloads/file/3535009/redirector-3.5.3-an+fx.xpi";
-          sha256 = "sha256-7dvT1ZROdI0L1uy22enPDgwC3O1vQtshqrZBkOccD3E=";
-          meta = {};
-        };
-
         tetrio-plus = buildFirefoxXpiAddon {
           pname = "tetrio-plus";
           version = "0.23.7";